MyPython_test/streamlit/supu_app/requirements.txt
Pillow==9.1.1
pandas==1.4.2
